Move marginal MAP to experimental package
This should be easy to revert if we want to move it back in e.g. Figaro
4.2.
