+++ Figaro/src/main/scala/com/cra/figaro/experimental/marginalmap/MarginalMAPVEStrategy.scala
@@ -10,13 +10,11 @@
  *
  * See http://www.github.com/p2t2/figaro for a copy of the software license.
  */
-package com.cra.figaro.algorithm.structured.strategy.solve
+package com.cra.figaro.experimental.marginalmap
 
-import com.cra.figaro.algorithm.structured.Problem
-import com.cra.figaro.algorithm.factored.factors.Factor
-import com.cra.figaro.algorithm.factored.factors.Variable
-import com.cra.figaro.algorithm.structured.solver
-import com.cra.figaro.algorithm.structured.NestedProblem
+import com.cra.figaro.algorithm.factored.factors.{Factor, Variable}
+import com.cra.figaro.algorithm.structured.{NestedProblem, Problem, solver}
+import com.cra.figaro.algorithm.structured.strategy.solve.SolvingStrategy
 
 /**
  * A solving strategy that uses MPE VE to solve non-nested problems, and performs the MAP step at the top level.
